@@ -64,14 +64,15 @@
 
     <!-- Project -->
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-    <java.version>1.8</java.version>
+    <java.version>11</java.version>
     <!-- Dependency versions -->
-    <jersey.version>2.27</jersey.version>
+    <jersey.version>3.0.0</jersey.version>
     <jetty.version>11.0.1</jetty.version>
     <log4j.version>2.11.0</log4j.version>
-    <servlet.version>3.1.0</servlet.version>
+    <servlet.version>5.0.0</servlet.version>
     <slf4j.version>1.7.28</slf4j.version>
     <junit-jupiter.version>5.5.1</junit-jupiter.version>
+    <jakarta.xml.bind-api.version>3.0.0</jakarta.xml.bind-api.version>
   </properties>
 
   <modules>
@@ -84,9 +85,14 @@
   <dependencyManagement>
     <dependencies>
       <dependency>
-        <groupId>javax.servlet</groupId>
-        <artifactId>javax.servlet-api</artifactId>
+        <groupId>jakarta.servlet</groupId>
+        <artifactId>jakarta.servlet-api</artifactId>
         <version>${servlet.version}</version>
+      </dependency>
+      <dependency>
+        <groupId>jakarta.xml.bind</groupId>
+        <artifactId>jakarta.xml.bind-api</artifactId>
+        <version>${jakarta.xml.bind-api.version}</version>
       </dependency>
       <dependency>
         <groupId>org.slf4j</groupId>
@@ -199,10 +205,10 @@
       <plugin>
         <groupId>org.apache.maven.plugins</groupId>
         <artifactId>maven-compiler-plugin</artifactId>
-        <version>3.7.0</version>
+        <version>3.8.1</version>
         <configuration>
-          <source>1.8</source>
-          <target>1.8</target>
+          <source>${java.version}</source>
+          <target>${java.version}</target>
           <encoding>${project.build.sourceEncoding}</encoding>
         </configuration>
       </plugin>
